--- RateLimiter.Api/Infrastructure/Authentication/JwtBearerExtensions.cs
@@ -1,8 +1,0 @@
-﻿namespace RateLimiter.Api.Infrastructure.Authentication
-{
-	public static class JwtBearerExtensions
-	{
-		public static void AddJwtBearerConfiguration(this IServiceCollection services)
-			=> services.ConfigureOptions<ConfigureJwtBearerOptions>();
-	}
-}
